+++ packages/identity-cli/src/commands/verifiableCredentialUnrevoke.ts
@@ -1,12 +1,12 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { CLIDisplay, CLIParam } from "@gtsc/cli-core";
-import { Converter, I18n } from "@gtsc/core";
-import { IotaIdentityConnector } from "@gtsc/identity-connector-iota";
-import { VaultConnectorFactory } from "@gtsc/vault-models";
+import { CLIDisplay, CLIParam } from "@twin.org/cli-core";
+import { Converter, I18n } from "@twin.org/core";
+import { IotaIdentityConnector } from "@twin.org/identity-connector-iota";
+import { VaultConnectorFactory } from "@twin.org/vault-models";
 
-import { IotaWalletConnector } from "@gtsc/wallet-connector-iota";
-import { WalletConnectorFactory } from "@gtsc/wallet-models";
+import { IotaWalletConnector } from "@twin.org/wallet-connector-iota";
+import { WalletConnectorFactory } from "@twin.org/wallet-models";
 import { Command } from "commander";
 import { setupVault } from "./setupCommands";