+++ packages/identity-cli/src/commands/verificationMethodRemove.ts
@@ -1,13 +1,13 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { CLIDisplay, CLIOptions, CLIParam } from "@gtsc/cli-core";
-import { Converter, I18n, StringHelper } from "@gtsc/core";
-import { IotaIdentityConnector, IotaIdentityUtils } from "@gtsc/identity-connector-iota";
-import { DocumentHelper } from "@gtsc/identity-models";
-import { VaultConnectorFactory } from "@gtsc/vault-models";
+import { CLIDisplay, CLIOptions, CLIParam } from "@twin.org/cli-core";
+import { Converter, I18n, StringHelper } from "@twin.org/core";
+import { IotaIdentityConnector, IotaIdentityUtils } from "@twin.org/identity-connector-iota";
+import { DocumentHelper } from "@twin.org/identity-models";
+import { VaultConnectorFactory } from "@twin.org/vault-models";
 
-import { IotaWalletConnector } from "@gtsc/wallet-connector-iota";
-import { WalletConnectorFactory } from "@gtsc/wallet-models";
+import { IotaWalletConnector } from "@twin.org/wallet-connector-iota";
+import { WalletConnectorFactory } from "@twin.org/wallet-models";
 import { Command } from "commander";
 import { setupVault } from "./setupCommands";