+++ packages/identity-models/src/models/IIdentityComponent.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IComponent } from "@gtsc/core";
-import type { IDidDocument } from "@gtsc/standards-w3c-did";
+import type { IComponent } from "@twin.org/core";
+import type { IDidDocument } from "@twin.org/standards-w3c-did";
 
 /**
  * Interface describing a contract which provides identity operations.