+++ packages/identity-models/src/models/IIdentityProfileConnector.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IComponent } from "@gtsc/core";
-import type { IJsonLdDocument } from "@gtsc/data-json-ld";
+import type { IComponent } from "@twin.org/core";
+import type { IJsonLdDocument } from "@twin.org/data-json-ld";
 
 /**
  * Interface describing a contract which provides profile operations.