+++ packages/identity-models/src/models/api/IIdentityProfileGetPublicResponse.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IJsonLdDocument } from "@gtsc/data-json-ld";
-import type { HeaderTypes, MimeTypes } from "@gtsc/web";
+import type { IJsonLdDocument } from "@twin.org/data-json-ld";
+import type { HeaderTypes, MimeTypes } from "@twin.org/web";
 
 /**
  * Response to get an identity public profile.