+++ packages/identity-models/src/utils/documentHelper.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { Is } from "@gtsc/core";
+import { Is } from "@twin.org/core";
 
 /**
  * Helper methods for documents.