+++ packages/identity-connector-iota/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/identity-connector-iota - Changelog
+# @twin.org/identity-connector-iota - Changelog
 
-## v0.0.1
+## v0.0.1-next.3
 
 - Initial Release